Conflicts:
	Classy.Models/Search/ListingIndexDto.cs
	Classy.UtilRunner/Utilities/Indexing/ProductIndexer.cs

+++ Classy.Repository/IShippingCalculator.cs
@@ -1,17 +1,9 @@
 ﻿using Classy.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Classy.Repository
 {
     public interface IShippingCalculator
     {
-        double GetShippingPrice(
-            Profile merchantProfile, 
-            Location origin, 
-            PhysicalAddress destination);
+        decimal GetShippingPrice(Profile merchantProfile, Location origin, PhysicalAddress destination);
     }
 }
